Refactor : refactor the proxy cluster container for suite and scaling
