+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-integration-test-scaling/src/test/java/org/apache/shardingsphere/integration/data/pipeline/framework/container/config/proxy/ScalingProxyClusterContainerConfiguration.java
@@ -1,3 +1,20 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.apache.shardingsphere.integration.data.pipeline.framework.container.config.proxy;
 
 import lombok.RequiredArgsConstructor;
@@ -13,7 +30,7 @@ import java.util.Map;
  * Scaling proxy cluster container configuration.
  */
 @RequiredArgsConstructor
-public class ScalingProxyClusterContainerConfiguration implements AdaptorContainerConfiguration {
+public final class ScalingProxyClusterContainerConfiguration implements AdaptorContainerConfiguration {
     
     private final DatabaseType databaseType;
     
@@ -25,7 +42,7 @@ public class ScalingProxyClusterContainerConfiguration implements AdaptorContain
     }
     
     @Override
-    public Map<String, String> getResourceMappings(String scenario, DatabaseType databaseType) {
+    public Map<String, String> getResourceMappings(final String scenario, final DatabaseType databaseType) {
         Map<String, String> result = new HashMap<>(2);
         result.put("/env/logback.xml", "/opt/shardingsphere-proxy/conf/logback.xml");
         if (DatabaseTypeUtil.isMySQL(databaseType)) {
