+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-integration-test-suite/src/test/java/org/apache/shardingsphere/test/integration/container/compose/mode/ClusterComposedContainer.java
@@ -17,7 +17,9 @@
 
 package org.apache.shardingsphere.test.integration.container.compose.mode;
 
+import org.apache.shardingsphere.test.integration.container.compose.ComposedContainer;
 import org.apache.shardingsphere.test.integration.container.config.SuiteStorageContainerConfigurationFactory;
+import org.apache.shardingsphere.test.integration.container.config.proxy.SuiteProxyClusterContainerConfiguration;
 import org.apache.shardingsphere.test.integration.env.container.atomic.DockerITContainer;
 import org.apache.shardingsphere.test.integration.env.container.atomic.ITContainers;
 import org.apache.shardingsphere.test.integration.env.container.atomic.adapter.AdapterContainer;
@@ -26,7 +28,6 @@ import org.apache.shardingsphere.test.integration.env.container.atomic.governanc
 import org.apache.shardingsphere.test.integration.env.container.atomic.governance.GovernanceContainerFactory;
 import org.apache.shardingsphere.test.integration.env.container.atomic.storage.StorageContainer;
 import org.apache.shardingsphere.test.integration.env.container.atomic.storage.StorageContainerFactory;
-import org.apache.shardingsphere.test.integration.container.compose.ComposedContainer;
 import org.apache.shardingsphere.test.integration.framework.param.model.ParameterizedArray;
 
 import javax.sql.DataSource;
@@ -53,8 +54,9 @@ public final class ClusterComposedContainer implements ComposedContainer {
         // TODO add more version of databases
         storageContainer = containers.registerContainer(StorageContainerFactory.newInstance(parameterizedArray.getDatabaseType(), "", scenario,
                 SuiteStorageContainerConfigurationFactory.newInstance(parameterizedArray.getDatabaseType(), scenario)));
+        SuiteProxyClusterContainerConfiguration suiteProxyClusterContainerConfiguration = new SuiteProxyClusterContainerConfiguration(scenario);
         AdapterContainer adapterContainer = AdapterContainerFactory.newInstance(
-                parameterizedArray.getMode(), parameterizedArray.getAdapter(), parameterizedArray.getDatabaseType(), storageContainer, scenario);
+                parameterizedArray.getMode(), parameterizedArray.getAdapter(), parameterizedArray.getDatabaseType(), storageContainer, scenario, suiteProxyClusterContainerConfiguration);
         if (adapterContainer instanceof DockerITContainer) {
             ((DockerITContainer) adapterContainer).dependsOn(governanceContainer, storageContainer);
         }
